requirements.txt
absl-py==2.1.0
attrs==22.1.0
audtorch==0.6.4
biopython==1.83
biotite==0.40.0
debugpy==1.6.3
dgl==1.1.0
fair-esm @ git+https://github.com/facebookresearch/esm.git@900251ba3e2b7cdc06b44b10dfa3a0c1dd49752b
gpustat==1.1.1
huggingface-hub==0.23.2
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
matplotlib
ml_collections==0.1.1
mol2vec==0.2.2
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
OpenMM
packaging==24.0
pandas==2.2.2
pillow==10.2.0
pkgtools==0.7.3
protobuf==3.19.0
pyg==0.7.1
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
torch==2.1.2
torch_geometric==2.4.0
torchaudio==2.1.2
torchdata==0.7.1
torchinfo==1.8.0
torchvision==0.16.2
tqdm==4.64.1
transformers==4.41.1
tzdata==2024.1
urllib3==1.26.13
zipp==3.18.1
